pyproject.toml
[project]
name = "funniest-joke-with-llms"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=3.6.0",
    "dotenv>=0.9.9",
    "faiss-cpu>=1.11.0",
    "groq>=0.26.0",
    "nltk>=3.9.1",
    "numpy>=2.2.6",
    "pytest>=8.4.0",
    "sentence-transformers>=4.1.0",
    "typing>=3.10.0.0",
]


[tool.pytest.ini_options]
pythonpath = ["src"]
